requirements.txt
setuptools
fastapi==0.103.1
matplotlib==3.5.3
numpy==1.23.5
Pillow==9.1.1
requests
uvicorn
protobuf==3.19.6
python-multipart
